refactor(futures): decompose futures_data.py god module — PR 2/2 (#199)
- Extract futures_contracts.py (~130 lines): SYMBOL_EXCHANGE_MAP,
  _resolve_exchange, MAX_HISTORY_YEARS, FuturesContractService
- Extract futures_aggregates.py (~350 lines): FuturesAggregatesService
  with _download_contract, _download_full_history, _fill_data_gaps
- Extract futures_rollovers.py (~155 lines): FuturesRolloversService,
  _detect_single_rollover, _build_time_slices, CALENDAR_ROLL_DAYS_BEFORE_EXPIRY
- Extract futures_series.py (~115 lines): FutureSeriesService with
  get_continuous_series and _get_continuous_series_with_db
- Reduce futures_data.py from 1,086 → 87 lines as explicit-delegation facade
- sync_contracts/get_continuous_series use explicit method bodies to preserve
  patch("app.services.futures_data.SessionLocal", ...) test seams
- Private methods assigned via staticmethod() — no session management

All 9 import sites unchanged. All existing futures tests pass.
